fix: [#256] cancer and multiassay support pull_sheets.py
